Creates the schedule for the awards script
Also Resolves #283
